+++ include/InteractiveToolkit/MathCore/impl/quat_all_no_optimization.h
@@ -24,7 +24,7 @@ namespace MathCore
 
     public:
 
-        static constexpr size_t array_count = 4;
+        static constexpr int array_count = 4;
         using type = self_type;
 		using element_type = _BaseType;
