+++ include/InteractiveToolkit/MathCore/impl/vec2_float_simd.h
@@ -46,7 +46,7 @@ namespace MathCore
 
     public:
 
-        static constexpr size_t array_count = 2;
+        static constexpr int array_count = 2;
         using type = self_type;
 		using element_type = _BaseType;
