+++ include/InteractiveToolkit/MathCore/impl/vec4_float_simd.h
@@ -51,7 +51,7 @@ namespace MathCore
 
 	public:
 
-		static constexpr size_t array_count = 4;
+		static constexpr int array_count = 4;
 		using type = self_type;
 		using element_type = _BaseType;
